refactor: Update CMakeLists.txt to refine source file inclusion and add new unit tests
- Excluded test files and helpers from library source files in CMakeLists.txt.
- Added unit tests for serialClose, serialOpen, serialRead, and serialWrite functions.
- Introduced error capture mechanism for better error handling in tests.
- Implemented integration tests for serial communication with Arduino and pipe round trips.
- Removed outdated test_serial_functions.cpp file.
