+++ CMakeLists.txt
@@ -57,10 +57,10 @@ CPMAddPackage(
     "gtest_force_shared_crt ON"
 )
 
-file(
-    GLOB_RECURSE LIB_SOURCES
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp"
-)
+# Library sources: src/*.cpp only, exclude *.test.cpp and test_helpers/
+file(GLOB_RECURSE LIB_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp")
+list(FILTER LIB_SOURCES EXCLUDE REGEX ".*\\.test\\.cpp$")
+list(FILTER LIB_SOURCES EXCLUDE REGEX ".*/test_helpers/.*")
 
 add_library(cpp_bindings_linux SHARED ${LIB_SOURCES})
 
@@ -88,11 +88,11 @@ target_link_libraries(
 
 target_compile_features(cpp_bindings_linux PUBLIC cxx_std_23)
 
-# Collect all test source files
-file(
-    GLOB_RECURSE TEST_SOURCES
-    "${CMAKE_CURRENT_SOURCE_DIR}/tests/*.cpp"
-)
+# Test sources: src/*.test.cpp, tests/*.test.cpp, src/test_helpers/*.cpp (helpers excluded from lib)
+file(GLOB SRC_UNIT_TESTS "${CMAKE_CURRENT_SOURCE_DIR}/src/*.test.cpp")
+file(GLOB TESTS_INTEGRATION "${CMAKE_CURRENT_SOURCE_DIR}/tests/*.test.cpp")
+file(GLOB TEST_HELPER_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/src/test_helpers/*.cpp")
+set(TEST_SOURCES ${SRC_UNIT_TESTS} ${TESTS_INTEGRATION} ${TEST_HELPER_SOURCES})
 
 if(TEST_SOURCES)
     add_executable(cpp_bindings_linux_tests ${TEST_SOURCES})
@@ -101,6 +101,7 @@ if(TEST_SOURCES)
         cpp_bindings_linux_tests
         PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/include
+        ${CMAKE_CURRENT_SOURCE_DIR}/src
         ${CMAKE_BINARY_DIR}/generated
     )
