+++ tests/serial_arduino.test.cpp
@@ -1,4 +1,4 @@
-// Test for serial communication with Arduino echo script on /dev/ttyUSB0
+// Integration test: serial communication with Arduino echo script on /dev/ttyUSB0
 
 #include <cpp_core/interface/serial_close.h>
 #include <cpp_core/interface/serial_open.h>
@@ -8,18 +8,20 @@
 #include <cpp_core/status_codes.h>
 #include <gtest/gtest.h>
 
+#include <array>
 #include <cstring>
+#include <unistd.h>
 
 class SerialArduinoTest : public ::testing::Test
 {
   protected:
     void SetUp() override
     {
-        const char *env_port = std::getenv("SERIAL_TEST_PORT");
+        const char *env_port = std::getenv("SERIAL_TEST_PORT"); // NOLINT(concurrency-mt-unsafe)
         const char *port = env_port != nullptr ? env_port : "/dev/ttyUSB0";
-        handle_ = serialOpen(const_cast<void *>(static_cast<const void *>(port)), 115200, 8, 0, 0, nullptr);
+        handle = serialOpen(const_cast<void *>(static_cast<const void *>(port)), 115200, 8, 0, 0, nullptr);
 
-        if (handle_ <= 0)
+        if (handle <= 0)
         {
             GTEST_SKIP() << "Could not open serial port '" << port
                          << "'. Set SERIAL_TEST_PORT or connect Arduino on /dev/ttyUSB0.";
@@ -30,19 +32,19 @@ class SerialArduinoTest : public ::testing::Test
 
     void TearDown() override
     {
-        if (handle_ > 0)
+        if (handle > 0)
         {
-            serialClose(handle_, nullptr);
-            handle_ = 0;
+            serialClose(handle, nullptr);
+            handle = 0;
         }
     }
 
-    intptr_t handle_ = 0;
+    intptr_t handle = 0;
 };
 
 TEST_F(SerialArduinoTest, OpenClose)
 {
-    EXPECT_GT(handle_, 0) << "serialOpen should return a positive handle";
+    EXPECT_GT(handle, 0) << "serialOpen should return a positive handle";
 }
 
 TEST_F(SerialArduinoTest, WriteReadEcho)
@@ -50,52 +52,52 @@ TEST_F(SerialArduinoTest, WriteReadEcho)
     const char *test_message = "Hello Arduino!\n";
     int message_len = static_cast<int>(strlen(test_message));
 
-    int written = serialWrite(handle_, test_message, message_len, 1000, 1, nullptr);
+    int written = serialWrite(handle, test_message, message_len, 1000, 1, nullptr);
     EXPECT_EQ(written, message_len) << "Should write all bytes. Written: " << written << ", Expected: " << message_len;
 
     usleep(500000);
 
-    char read_buffer[256] = {0};
-    int read_bytes = serialRead(handle_, read_buffer, sizeof(read_buffer) - 1, 2000, 1, nullptr);
+    std::array<char, 256> read_buffer{};
+    int read_bytes = serialRead(handle, read_buffer.data(), static_cast<int>(read_buffer.size()) - 1, 2000, 1, nullptr);
 
     EXPECT_GT(read_bytes, 0) << "Should read at least some bytes";
-    EXPECT_LE(read_bytes, static_cast<int>(sizeof(read_buffer) - 1)) << "Should not overflow buffer";
+    EXPECT_LE(read_bytes, static_cast<int>(read_buffer.size()) - 1) << "Should not overflow buffer";
 
-    read_buffer[read_bytes] = '\0';
-    EXPECT_STRNE(read_buffer, "") << "Should receive echo from Arduino";
+    read_buffer[static_cast<size_t>(read_bytes)] = '\0';
+    EXPECT_STRNE(read_buffer.data(), "") << "Should receive echo from Arduino";
 }
 
 TEST_F(SerialArduinoTest, MultipleEchoCycles)
 {
-    const char *messages[] = {"Test1\n", "Test2\n", "Test3\n"};
-    const int num_messages = 3;
+    const std::array<const char *, 3> messages = {"Test1\n", "Test2\n", "Test3\n"};
 
-    for (int i = 0; i < num_messages; ++i)
+    for (size_t i = 0; i < messages.size(); ++i)
     {
         int msg_len = static_cast<int>(strlen(messages[i]));
 
-        int written = serialWrite(handle_, messages[i], msg_len, 1000, 1, nullptr);
+        int written = serialWrite(handle, messages[i], msg_len, 1000, 1, nullptr);
         EXPECT_EQ(written, msg_len) << "Cycle " << i << ": write failed";
 
         usleep(500000);
 
-        char read_buffer[256] = {0};
-        int read_bytes = serialRead(handle_, read_buffer, sizeof(read_buffer) - 1, 2000, 1, nullptr);
+        std::array<char, 256> read_buffer{};
+        int read_bytes =
+            serialRead(handle, read_buffer.data(), static_cast<int>(read_buffer.size()) - 1, 2000, 1, nullptr);
         EXPECT_GT(read_bytes, 0) << "Cycle " << i << ": read failed";
     }
 }
 
 TEST_F(SerialArduinoTest, ReadTimeout)
 {
-    char buffer[256];
-    int read_bytes = serialRead(handle_, buffer, sizeof(buffer), 100, 1, nullptr);
+    std::array<char, 256> buffer{};
+    int read_bytes = serialRead(handle, buffer.data(), static_cast<int>(buffer.size()), 100, 1, nullptr);
     EXPECT_GE(read_bytes, 0) << "Timeout should return 0, not error";
 }
 
 TEST(SerialInvalidHandleTest, InvalidHandleRead)
 {
-    char buffer[256];
-    int result = serialRead(-1, buffer, sizeof(buffer), 1000, 1, nullptr);
+    std::array<char, 256> buffer{};
+    int result = serialRead(-1, buffer.data(), static_cast<int>(buffer.size()), 1000, 1, nullptr);
     EXPECT_EQ(result, static_cast<int>(cpp_core::StatusCodes::kInvalidHandleError))
         << "Should return error for invalid handle";
 }
